+++ src/main/java/com/github/idea/ginkgo/execution/testing/GinkgoBuildRunningState.java
@@ -1,12 +1,9 @@
 package com.github.idea.ginkgo.execution.testing;
 
-import com.github.idea.ginkgo.GinkgoRunConfiguration;
 import com.github.idea.ginkgo.GinkgoRunConfigurationOptions;
 import com.goide.GoEnvironmentUtil;
 import com.goide.GoOsManager;
-import com.goide.execution.extension.GoExecutorExtension;
 import com.goide.i18n.GoBundle;
-import com.goide.sdk.GoSdkService;
 import com.goide.sdk.GoSdkUtil;
 import com.goide.util.GoUtil;
 import com.intellij.execution.DefaultExecutionResult;
@@ -14,18 +11,14 @@ import com.intellij.execution.ExecutionException;
 import com.intellij.execution.ExecutionResult;
 import com.intellij.execution.Executor;
 import com.intellij.execution.configurations.GeneralCommandLine;
-import com.intellij.execution.configurations.RunProfileState;
 import com.intellij.execution.process.*;
 import com.intellij.execution.runners.ExecutionEnvironment;
 import com.intellij.execution.runners.ProgramRunner;
 import com.intellij.execution.ui.RunContentDescriptor;
-import com.intellij.openapi.project.Project;
 import com.intellij.openapi.util.Couple;
 import com.intellij.openapi.util.io.FileUtil;
-import com.intellij.openapi.util.text.StringUtil;
 import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.util.EnvironmentUtil;
-import com.intellij.util.containers.ContainerUtil;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.jetbrains.concurrency.AsyncPromise;
@@ -33,29 +26,25 @@ import org.jetbrains.concurrency.AsyncPromise;
 import java.io.File;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
 
-public class GinkgoBuildRunningState implements RunProfileState {
-    private final ExecutionEnvironment environment;
-    private final Project project;
-    private final GinkgoRunConfiguration configuration;
+public class GinkgoBuildRunningState extends GinkgoState {
+    public static final String COMPILATION_FAILED = "go.execution.compilation.failed.notification.title";
+    public static final String FILE_CREATION_FAILED = "go.execution.cannot.create.temp.output.file.error";
     private final AsyncPromise<RunContentDescriptor> buildingPromise;
-    private VirtualFile sdkRoot;
     private File outputFile;
     private String buildCommand;
 
-
-    public GinkgoBuildRunningState(@NotNull ExecutionEnvironment env, @Nullable Project project, @NotNull GinkgoRunConfiguration configuration, AsyncPromise<RunContentDescriptor> buildingPromise) {
-        this.environment = env;
-        this.project = project;
-        this.configuration = configuration;
+    public GinkgoBuildRunningState(@NotNull ExecutionEnvironment env, AsyncPromise<RunContentDescriptor> buildingPromise, GinkgoRunningState runningState) {
+        super(env, runningState.getProject(), runningState.getConfiguration());
         this.buildingPromise = buildingPromise;
-        this.sdkRoot = GoSdkService.getInstance(project).getSdk(null).getSdkRoot();
     }
 
     @Override
     @Nullable
-    public ExecutionResult execute(Executor executor, ProgramRunner<?> runner) throws ExecutionException {
+    public ExecutionResult execute(Executor executor, @NotNull ProgramRunner<?> runner) throws ExecutionException {
         return new DefaultExecutionResult(startBuildProcess());
     }
 
@@ -81,7 +70,7 @@ public class GinkgoBuildRunningState implements RunProfileState {
             @Override
             public void processTerminated(@NotNull ProcessEvent event) {
                 if (event.getExitCode() != 0) {
-                    buildingPromise.setError(new ExecutionException(GoBundle.message("go.execution.compilation.failed.notification.title", new Object[0])));
+                    buildingPromise.setError(new ExecutionException(GoBundle.message(COMPILATION_FAILED)));
                 }
                 buildingPromise.setResult(environment.getContentToReuse());
             }
@@ -91,9 +80,9 @@ public class GinkgoBuildRunningState implements RunProfileState {
     }
 
     /**
-     * @param packageName   Name of the go package to be compiled for test
-     * @return              Temp compiled test file target
-     * @throws ExecutionException
+     * @param packageName Name of the go package to be compiled for test
+     * @return Temp compiled test file target
+     * @throws ExecutionException when can't create temporary file
      */
     private @NotNull File getOutputFile(@NotNull String packageName) throws ExecutionException {
         String binaryName = GoEnvironmentUtil.getBinaryFileNameForPath(FileUtil.sanitizeFileName(packageName), ".test", GoOsManager.isWindows());
@@ -101,46 +90,11 @@ public class GinkgoBuildRunningState implements RunProfileState {
         try {
             return FileUtil.createTempFile(outputDirectory, "", binaryName, true);
         } catch (IOException ioException) {
-            throw new ExecutionException(GoBundle.message("go.execution.cannot.create.temp.output.file.error", new Object[0]), ioException);
+            throw new ExecutionException(GoBundle.message(FILE_CREATION_FAILED), ioException);
         }
     }
 
-    /**
-     * Uses the Goland executor extension to get Go module environment variables from the project settings setup
-     * under Go -> Modules -> Environment
-     *
-     * @return Map<String, String> of Environmental entries from project configuration
-     */
-    @NotNull
-    private Map<String, String> getProjectEnvironmentExtensions() {
-        Map<String, String> environmentFromExtensions = new HashMap<>();
-        Iterator goExtensionIterator = GoExecutorExtension.EP_NAME.getExtensionList().iterator();
-
-        while (goExtensionIterator.hasNext()) {
-            GoExecutorExtension extension = (GoExecutorExtension) goExtensionIterator.next();
-            environmentFromExtensions.putAll(extension.getExtraEnvironment(project, null, environmentFromExtensions));
-        }
-        return environmentFromExtensions;
-    }
-
-    /**
-     * Updates the environment path with the go bin paths as determined by framework configuration.
-     *
-     * @param env
-     * @return Couple<String>
-     */
-    private Couple<String> updatePath(Map<String, String> env) {
-        Collection<String> paths = new ArrayList<>();
-        String goBinPaths = GoSdkUtil.retrieveEnvironmentPathForGo(project, null);
-        Couple<String> pathEntry = GoEnvironmentUtil.getPathEntry(env);
-
-        ContainerUtil.addIfNotNull(paths, StringUtil.nullize(goBinPaths, true));
-        ContainerUtil.addIfNotNull(paths, StringUtil.nullize(pathEntry.second, true));
-
-        return new Couple<>(pathEntry.first, StringUtil.join(paths, File.pathSeparator));
-    }
-
-    public GeneralCommandLine createBuildCommandLine() throws ExecutionException {
+    private GeneralCommandLine createBuildCommandLine() throws ExecutionException {
         GinkgoRunConfigurationOptions options = configuration.getOptions();
         outputFile = getOutputFile(options.getPackageName());
         String defaultExecutable = GoEnvironmentUtil.getBinaryFileNameForPath("go");
@@ -158,7 +112,7 @@ public class GinkgoBuildRunningState implements RunProfileState {
         commandList.add("all=-N -l");
         commandList.add(".");
 
-        return new GeneralCommandLine(commandList.stream().toArray(String[]::new));
+        return new GeneralCommandLine(commandList.toArray(new String[0]));
     }
 
     public File getOutputFile() {