+++ src/main/java/com/github/idea/ginkgo/execution/testing/GinkgoRunningState.java
@@ -4,18 +4,13 @@ import com.github.idea.ginkgo.GinkgoConsoleProperties;
 import com.github.idea.ginkgo.GinkgoRunConfiguration;
 import com.github.idea.ginkgo.GinkgoRunConfigurationOptions;
 import com.github.idea.ginkgo.scope.GinkgoScope;
-import com.goide.GoEnvironmentUtil;
 import com.goide.dlv.DlvVm;
 import com.goide.execution.GoRunUtil;
-import com.goide.execution.extension.GoExecutorExtension;
-import com.goide.sdk.GoSdkService;
-import com.goide.sdk.GoSdkUtil;
 import com.intellij.execution.DefaultExecutionResult;
 import com.intellij.execution.ExecutionException;
 import com.intellij.execution.ExecutionResult;
 import com.intellij.execution.Executor;
 import com.intellij.execution.configurations.GeneralCommandLine;
-import com.intellij.execution.configurations.RunProfileState;
 import com.intellij.execution.process.KillableColoredProcessHandler;
 import com.intellij.execution.process.ProcessHandler;
 import com.intellij.execution.process.ProcessOutputTypes;
@@ -28,10 +23,7 @@ import com.intellij.execution.testframework.sm.SMTestRunnerConnectionUtil;
 import com.intellij.execution.testframework.sm.runner.ui.SMTRunnerConsoleView;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.util.Couple;
-import com.intellij.openapi.util.text.StringUtil;
-import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.util.EnvironmentUtil;
-import com.intellij.util.containers.ContainerUtil;
 import com.intellij.util.net.NetUtils;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -46,12 +38,8 @@ import java.net.UnknownHostException;
 import java.nio.charset.StandardCharsets;
 import java.util.*;
 
-public class GinkgoRunningState implements RunProfileState {
-    private final ExecutionEnvironment environment;
-    private final Project project;
-    private final GinkgoRunConfiguration configuration;
+public class GinkgoRunningState extends GinkgoState {
     private InetSocketAddress myDebugAddress;
-    private VirtualFile sdkRoot;
     private File outputFile;
     private String buildCommand;
     @Nullable
@@ -59,23 +47,20 @@ public class GinkgoRunningState implements RunProfileState {
 
 
     public GinkgoRunningState(@NotNull ExecutionEnvironment env, @Nullable Project project, @NotNull GinkgoRunConfiguration configuration) {
-        this.environment = env;
-        this.project = project;
-        this.configuration = configuration;
-        this.sdkRoot = GoSdkService.getInstance(project).getSdk(null).getSdkRoot();
+        super(env, project, configuration);
     }
 
     @Override
     @Nullable
     public ExecutionResult execute(Executor executor, @NotNull ProgramRunner<?> runner) throws ExecutionException {
         if (executor.getId().equals("Debug")) {
-            return execute(executor, runner, startDebugProcess());
+            return execute(executor, startDebugProcess());
         }
-        return execute(executor, runner, startProcess());
+        return execute(executor, startProcess());
     }
 
     @NotNull
-    public ExecutionResult execute(@NotNull Executor executor, @NotNull ProgramRunner<?> runner, @NotNull ProcessHandler processHandler) {
+    public ExecutionResult execute(@NotNull Executor executor, @NotNull ProcessHandler processHandler) {
         GinkgoConsoleProperties consoleProperties = new GinkgoConsoleProperties(configuration, "Ginkgo", executor);
         SMTRunnerConsoleView consoleView = new SMTRunnerConsoleView(consoleProperties);
         SMTestRunnerConnectionUtil.initConsoleView(consoleView, "Ginkgo");
@@ -87,7 +72,6 @@ public class GinkgoRunningState implements RunProfileState {
         }
         return defaultExecutionResult;
     }
-
 
     @NotNull
     public ProcessHandler startProcess() throws ExecutionException {
@@ -150,42 +134,6 @@ public class GinkgoRunningState implements RunProfileState {
         return processHandler;
     }
 
-    /**
-     * Uses the Goland executor extension to get Go module environment variables from the project settings setup
-     * under Go -> Modules -> Environment
-     *
-     * @return Map<String, String> of Environmental entries from project configuration
-     */
-    @NotNull
-    private Map<String, String> getProjectEnvironmentExtensions() {
-        Map<String, String> environmentFromExtensions = new HashMap<>();
-        Iterator goExtensionIterator = GoExecutorExtension.EP_NAME.getExtensionList().iterator();
-
-        while (goExtensionIterator.hasNext()) {
-            GoExecutorExtension extension = (GoExecutorExtension) goExtensionIterator.next();
-            environmentFromExtensions.putAll(extension.getExtraEnvironment(project, null, environmentFromExtensions));
-        }
-        return environmentFromExtensions;
-    }
-
-    /**
-     * Updates the environment path with the go bin paths as determined by framework configuration.
-     *
-     * @param env
-     * @return Couple<String>
-     */
-    private Couple<String> updatePath(Map<String, String> env) {
-        Collection<String> paths = new ArrayList<>();
-        String goBinPaths = GoSdkUtil.retrieveEnvironmentPathForGo(project, null);
-        Couple<String> pathEntry = GoEnvironmentUtil.getPathEntry(env);
-
-        ContainerUtil.addIfNotNull(paths, StringUtil.nullize(goBinPaths, true));
-        ContainerUtil.addIfNotNull(paths, StringUtil.nullize(pathEntry.second, true));
-
-        return new Couple<>(pathEntry.first, StringUtil.join(paths, File.pathSeparator));
-    }
-
-
     private GeneralCommandLine createCommandLine(GinkgoRunConfigurationOptions runOptions) {
         List<String> commandList = new ArrayList<>();
         commandList.add(runOptions.getGinkgoExecutable());
@@ -198,14 +146,9 @@ public class GinkgoRunningState implements RunProfileState {
         }
 
         switch (runOptions.getGinkgoScope()) {
-            case ALL:
-                commandList.add("-r");
-                break;
-            case FOCUS:
-                commandList.add(String.format("--focus=%s", runOptions.getFocusTestExpression()));
-                break;
-            default:
-                throw new IllegalStateException("Unexpected value: " + runOptions.getGinkgoScope());
+            case ALL -> commandList.add("-r");
+            case FOCUS -> commandList.add(String.format("--focus=%s", runOptions.getFocusTestExpression()));
+            default -> throw new IllegalStateException("Unexpected value: " + runOptions.getGinkgoScope());
         }
         return new GeneralCommandLine(commandList);
     }
@@ -228,13 +171,9 @@ public class GinkgoRunningState implements RunProfileState {
         commandList.addAll(options.getGinkgoAdditionalOptionsList());
 
         switch (options.getGinkgoScope()) {
-            case ALL:
-                throw new ExecutionException("Can not debug on test scope all");
-            case FOCUS:
-                commandList.add(String.format("-ginkgo.focus=%s", options.getFocusTestExpression()));
-                break;
-            default:
-                throw new IllegalStateException("Unexpected value: " + options.getGinkgoScope());
+            case ALL -> throw new ExecutionException("Can not debug on test scope all");
+            case FOCUS -> commandList.add(String.format("-ginkgo.focus=%s", options.getFocusTestExpression()));
+            default -> throw new IllegalStateException("Unexpected value: " + options.getGinkgoScope());
         }
         return new GeneralCommandLine(commandList);
     }
@@ -256,8 +195,7 @@ public class GinkgoRunningState implements RunProfileState {
         return environment.getExecutor().getId().equals("Debug");
     }
 
-    @NotNull
-    public InetSocketAddress getDebugServerAddress() {
+    public void getDebugServerAddress() {
         if (myDebugAddress == null) {
             try {
                 myDebugAddress = new InetSocketAddress(resolveLocalAddress(), NetUtils.findAvailableSocketPort());
@@ -267,8 +205,6 @@ public class GinkgoRunningState implements RunProfileState {
                 DlvVm.LOG.warn("Cannot find free port", ioException);
             }
         }
-
-        return myDebugAddress;
     }
 
     private InetAddress resolveLocalAddress() throws UnknownHostException {
@@ -284,16 +220,12 @@ public class GinkgoRunningState implements RunProfileState {
         return configuration;
     }
 
-    public File getOutputFile() {
-        return outputFile;
+    public boolean isPackageLevel() {
+        return configuration.getOptions().getGinkgoScope() != GinkgoScope.ALL;
     }
 
     public void setOutputFile(File outputFile) {
         this.outputFile = outputFile;
-    }
-
-    public String getBuildCommand() {
-        return buildCommand;
     }
 
     public void setBuildCommand(String buildCommand) {